requirements.txt
asgiref==3.7.2
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.4
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
sendgrid==6.10.0
